requirements.txt
streamlit
yfinance
pandas
plotly
python-dotenv
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.24.0
kaleido
ta
nest_asyncio
pillow>=10.0.0
datasets>=2.14.0
peft>=0.6.0
trl>=0.7.4
bitsandbytes>=0.41.0
tensorboard>=2.14.0
huggingface_hub>=0.19.0 